Use QString for Mode::id instead of int
We can't know for sure that backend will use number as mode id, we
can't even know that they will give us a id at all.

This problem was shown in the XRandR1.1 backend where we instead of
having one Mode per size@refresh we have one mode with multiple refresh
but only one id.

For example, mode:
1 1920x1080 30, 50, 60

Mode 1 will have 1 size, and 3 different refresh date, which is
impossible to map to KScreen::Config since our Modes only have one
refresh.

So now that the Mode::id a string we can do magic things like:
1-0 1920x1080 30
1-1 1920x1080 50
1-2 1920x1080 60

+++ backends/xrandr/xrandrconfig.cpp
@@ -369,7 +369,7 @@ bool XRandRConfig::enableOutput(Output* output) const
     RROutput *outputs = new RROutput[1];
     outputs[0] = output->id();
     Status s = XRRSetCrtcConfig(XRandR::display(), XRandR::screenResources(), XRandR::freeCrtc(output->id()),
-        CurrentTime, output->pos().rx(), output->pos().ry(), output->currentModeId(),
+        CurrentTime, output->pos().rx(), output->pos().ry(), output->currentModeId().toInt(),
         output->rotation(), outputs, 1);
 
     kDebug(dXndr()) << "XRRSetCrtcConfig() returned" << s;
@@ -384,7 +384,7 @@ bool XRandRConfig::changeOutput(Output* output, int crtcId) const
     RROutput *outputs = new RROutput[1];
     outputs[0] = output->id();
     Status s = XRRSetCrtcConfig(XRandR::display(), XRandR::screenResources(), crtcId,
-        CurrentTime, output->pos().rx(), output->pos().ry(), output->currentModeId(),
+        CurrentTime, output->pos().rx(), output->pos().ry(), output->currentModeId().toInt(),
         output->rotation(), outputs, 1);
 
     kDebug(dXndr()) << "XRRSetCrtcConfig() returned" << s;
